+++ src/model/entities/Boss1.java
@@ -16,8 +16,8 @@ public class Boss1 extends Enemy {
 
 	private final int INITIAL_WIDTH = 0;
 	private final int INITIAL_HEIGHT = 0;
-	private final int INITIAL_MU_X = 0;
-	private final int INITIAL_MU_Y = 0;
+	private final double INITIAL_MU_X = 0;
+	private final double INITIAL_MU_Y = 0;
 	private final int MAX_HITS = 0;
 
 	private EntityDirections direction;
